Closes #28. Redirect links in Physics and Chemistry Nobel Laureates. Closes #11. Remove Royal Prussia from the list of physicists.
1. Update the Nobel Laureates in Physics and Chemistry to have the names
of any redirected links. Regenerated all the data files downstream as a
consequence of this.
2. Added alphabetical sorting to fields when creating the dictionary for
the physicists so that comparison for testing is deterministic. The order
of several of the fields was messed up when regenerating the dataframe.
This caused the test to fail even though the data was the same.
3. Add Royal Prussia to the blacklist so it is not collected as a
physicist.
4. Added autopep8 for development which seems to have updated jupyter lab
version.