Closes #33. Proper quoting in URLs.
Always quote the filename of the URL by calling `parse.quote`
which knows whether to quote the text.

Also refactor to use the same method for fetching the URLs.

Note that my fix to the issue with new lines is incorrect as
I have now introduced some stars into the values. I will fix
this in due course.