Closes #34. Refactor code that fetches URLs to work asynchronously.
Major performance improvements by using `requests-futures` library
and refactoring the code to work asynchronously. Also fix the method
in 2.0-process-physicists-raw-data.ipynb that fetches the redirected
URLs to only fetch the URL if it is not in the cache.

- Also add revision IDs when fetching Wikipedia pages for
reproducibility.
- Also remove code that replaces \n* in 2.0-process-physicists-raw-data.ipynb
since it is not working correctly and causing parts of the abstract to
be removed. This is the reason some of the unit tests are failing. I
will fix this in due course.