Closes #48. Impute missing values for country codes using nationality list.
Add external data for mapping countries and denonyms to country codes.
Read this into a dataframe and append to the dataframe some other
common variants or ex countries that are often used. Use this dataframe to
map the places to their respective country codes. Persist both the places
and nationality list dataframe for later use in feature construction.