Closes #45. Build country of citizenship and continent codes features.
Create country of citizenship feature from a concatenation of the
citizenship, nationality and description fields. Also create features
for the continent codes and the number of each of these.

Move `nationality_to_alpha2_code` function to a separate utility class
as it is called from multiple notebooks. Refactor the method:

1. To not use named entity recognition for Northern Ireland.
2. Only initialize the denomyns list once by using a function attribute.

These were causing a severe performance issue.