Closes #53. Convert gender and target to string in build features
Convert gender to string (male / female) when building features.
This is so that the Factor Analysis for Mixed Data (FAMD) will
treat this variable as categorical during dimensionality
reduction.

Convert target to string (yes / no) as it has more clarity and
is easy enough to map to (1 / 0) during machine learning.

Also drop 'name' variable and rename 'fullName' to 'full_name'
as snake_case is nicer.