Closes #56. Build features for the test data.
Create features for the test set. Syncronize
the test set features with the training set features
due to the one hot encoding. Also joined the training
and test dataframes for the physicists when merging
the names for the Nobel Physics Laureates and Chemistry
Laureates list. This led to extra Nobel Laureates being
found in the feature building and hence the values for
3 physicists changed in the training features.