Closes #58. Convert all count features to ratio variables.
Convert all discrete count features to continuous ratio
variables. Had to refactor code so that I could use the mean
value of the training set as the denominator in the ratio for
both training and test sets. I also added some code to remove
features where all the counts were zero. This removed 3 columns
from the features dataframe related to Nobel Laureates counts.
Also, updated the column names to refer to ratios now as well
as the prose in the notebook.