Closes #60. Group binary features that are very rare.
Bucket binary features that are very rare into an "other" category
by using a presence threshold to indicate the fraction of physicists
in which the value must appear to not be grouped. This reduces the
dimensionality of the feature space which should help with trying to
prevent overfitting when building models.