Closes #62. Create a validation set from the test data.
Create a validation set using data from the test set
as there is a pseudo-time component to the data. This
results in a training, validation and test set. The
training set consists of dead physicists whilst the other
two sets consist of alive physicists. The validation and
test sets are split in two according to the date of birth.
Earlier dates of birth of birth are in the validation set
and later dates of birth in the test set.

Also fix bug where there are two dead physicists
without death dates in the test set. Impute the values
of death dates which increases the length of the training
data by two.