Closes #79. Train a support vector machine classifier.
Train a support vector machine classifier. Choose the
value of regularization paramater on the training error,
`C`, and the `class_weight` using the maximum Matthews
Correlation Coefficient on the validation set. Make sure
to fix the `kernel` hyperparamter as `linear` since this
is the only kernel for which weights of the separating
hyperplane are available.