Closes #81. Model selection.
Produce a plot of Matthews Correlation Coefficient (MCC) vs (binary)
classification threshold for each of the classifiers. For SVM transform
the decision function threshold to a value between 0 and 1 so it is on
the same scale as the other classifiers. Define and compute the
"normalized area under the MCC (NAUC MCC) curve" and select the model
with the highest value.