Closes #98. Cleanup collect-nobel-laureates.
Cleanup the narrative in this notebook. Also added Maria
Goeppert-Meyer to redirect list as Wikipedia has changed
this to Maria Goeppert Meyer. This causes a minor issue
for reproducibility which I should probably look at.
Updated the test to new API. Updated the git ignore file
for a new `pywikibot file`.